Fix sản phẩm

+++ project/src/main/java/com/repository/SanPhamDAO.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.model.SanPham;
 
 @Repository
-public interface SanPhamDAO  extends JpaRepository<SanPham, Long> {
+public interface SanPhamDAO  extends JpaRepository<SanPham, Integer> {
 
 }
